pyproject.toml
[project]
name = "pax-mcp"
version = "0.1.0"
description = "MCP server for scanning local servers for security vulnerabilities"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["mcp", "security", "vulnerability-scanner", "ngrok"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
  "fastmcp",
  "requests",
  "pydantic",
  "python-dotenv",
  "ngrok",
  "crawl4ai",
  "openai",
  "fastapi",
  "uvicorn",
  "aiohttp",
]

[project.scripts]
pax-mcp = "server:main"

[project.urls]
Homepage = "https://github.com/yourusername/pax-mcp"
Repository = "https://github.com/yourusername/pax-mcp"
Issues = "https://github.com/yourusername/pax-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
